requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
langchain_google_genai
faiss-cpu